+++ front/validate_login.py
@@ -1,10 +1,11 @@
 #!/usr/bin/python
 
-import cgi, Cookie, LINK_HEADERS, sys
+import cgi, Cookie, LINK_HEADERS, sys,os
 sys.path.insert(0 , str(LINK_HEADERS.DATABASE_LINK))
 from database_class import DB
 sys.path.insert(1 , str(LINK_HEADERS.SCRIPTS_LINK))
-from auth_class import Auth
+sys.path.insert(1 , str(LINK_HEADERS.SERVICE_LINK))
+from auth_service import Auth
 
 form = cgi.FieldStorage()
 username = str(form.getvalue("username"))
